# docs/api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Xenoteer Control API",
    "version": "1.0.0-phase4",
    "description": "Implemented Phase 4 HTTP control, observation, clipboard, capture, artifact, and viewer contract. Typed request and result definitions are generated from xenoteer-protocol and referenced from schemas/v1. Control WebSocket application messages are documented separately in websocket.md.",
    "license": {
      "name": "Apache License 2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Default container-local control plane; production TLS may terminate at a trusted same-host proxy."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/desktops/{desktop_id}/windows": {
      "parameters": [
        {"$ref": "#/components/parameters/DesktopId"}
      ],
      "get": {
        "operationId": "listWindows",
        "tags": ["Windows"],
        "summary": "List one bounded, atomic page of windows",
        "description": "Requires desktop:observe. The cursor is opaque and bound to the principal, generation, model revision, order, and query.",
        "parameters": [
          {"$ref": "#/components/parameters/DesktopGenerationQuery"},
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {"type": "integer", "format": "uint16", "minimum": 1, "maximum": 200, "default": 50}
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["creation_ascending", "creation_descending", "stacking_bottom_to_top", "stacking_top_to_bottom", "title_ascending", "title_descending", "xid_ascending"],
              "default": "creation_ascending"
            }
          },
          {"$ref": "#/components/parameters/WindowCursorQuery"}
        ],
        "responses": {
          "200": {
            "description": "Atomic window page. Cache-Control is private, no-store.",
            "content": {"application/json": {"schema": {"$ref": "../../../schemas/v1/window-list-page.json"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "429": {"$ref": "#/components/responses/TooManyRequests"},
          "503": {"$ref": "#/components/responses/CapabilityUnavailable"},
          "504": {"$ref": "#/components/responses/RequestOutcomeUnknown"}
        }
      }
    },
    "/v1/desktops/{desktop_id}/windows/query": {
      "parameters": [{"$ref": "#/components/parameters/DesktopId"}],
      "post": {
        "operationId": "queryWindows",
        "tags": ["Windows"],
        "summary": "Evaluate a bounded selector against one model revision",
        "description": "Requires desktop:observe. Path and body desktop IDs must match.",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "../../../schemas/v1/window-query-request.json"}}}
        },
        "responses": {
          "200": {
            "description": "Atomic selector result page. Cache-Control is private, no-store.",
            "content": {"application/json": {"schema": {"$ref": "../../../schemas/v1/window-query-page.json"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "413": {"$ref": "#/components/responses/PayloadTooLarge"},
          "429": {"$ref": "#/components/responses/TooManyRequests"},
          "503": {"$ref": "#/components/responses/CapabilityUnavailable"},
          "504": {"$ref": "#/components/responses/RequestOutcomeUnknown"}
        }
      }
    },
    "/v1/desktops/{desktop_id}/windows/resolve": {
      "parameters": [{"$ref": "#/components/parameters/DesktopId"}],
      "post": {
        "operationId": "resolveWindow",
        "tags": ["Windows"],
        "summary": "Resolve a selector to one exact live window birth",
        "description": "Requires desktop:observe. Ambiguous selectors fail instead of choosing an arbitrary target.",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "../../../schemas/v1/window-resolve-request.json"}}}
        },
        "responses": {
          "200": {
            "description": "Exact resolved window and opaque reference token. Cache-Control is private, no-store.",
            "content": {"application/json": {"schema": {"$ref": "../../../schemas/v1/window-resolve-result.json"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "413": {"$ref": "#/components/responses/PayloadTooLarge"},
          "429": {"$ref": "#/components/responses/TooManyRequests"},
          "503": {"$ref": "#/components/responses/CapabilityUnavailable"},
          "504": {"$ref": "#/components/responses/RequestOutcomeUnknown"}
        }
      }
    },
    "/v1/desktops/{desktop_id}/windows/wait": {
      "parameters": [{"$ref": "#/components/parameters/DesktopId"}],
      "post": {
        "operationId": "waitWindow",
        "tags": ["Windows"],
        "summary": "Wait for one bounded window predicate",
        "description": "Requires desktop:observe. The model actor atomically checks, registers, and rechecks the predicate. Admission is bounded; timeout and resynchronization are represented in the 200 response body.",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "../../../schemas/v1/window-wait-request.json"}}}
        },
        "responses": {
          "200": {
            "description": "Satisfied, timed-out, or resynchronization-required wait result. Cache-Control is private, no-store.",
            "content": {"application/json": {"schema": {"$ref": "../../../schemas/v1/window-wait-result.json"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "413": {"$ref": "#/components/responses/PayloadTooLarge"},
          "429": {"$ref": "#/components/responses/TooManyRequests"},
          "503": {"$ref": "#/components/responses/CapabilityUnavailable"},
          "504": {"$ref": "#/components/responses/RequestOutcomeUnknown"}
        }
      }
    },
    "/v1/desktops/{desktop_id}/windows/{reference_token}": {
      "parameters": [
        {"$ref": "#/components/parameters/DesktopId"},
        {"$ref": "#/components/parameters/WindowReferenceToken"}
      ],
      "get": {
        "operationId": "getWindowSnapshot",
        "tags": ["Windows"],
        "summary": "Resolve an opaque token to one exact live window snapshot",
        "description": "Requires desktop:observe. Tokens are principal-, generation-, and exact-birth-bound and expire.",
        "parameters": [{"$ref": "#/components/parameters/DesktopGenerationQuery"}],
        "responses": {
          "200": {
            "description": "Exact snapshot and model revision. Cache-Control is private, no-store.",
            "content": {"application/json": {"schema": {"$ref": "../../../schemas/v1/window-snapshot-result.json"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "429": {"$ref": "#/components/responses/TooManyRequests"},
          "503": {"$ref": "#/components/responses/CapabilityUnavailable"},
          "504": {"$ref": "#/components/responses/RequestOutcomeUnknown"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Exactly one long-lived API Bearer credential. The token is opaque and must not be placed in URLs. The closed grant vocabulary is desktop:status, desktop:observe, input:control, application:launch, application:terminate, window:control, clipboard:read, clipboard:write, capture:read, artifact:read, artifact:delete, and viewer:read. This scheme is not accepted by the viewer gateway WebSocket."
      },
      "viewerTicketSubprotocol": {
        "type": "apiKey",
        "in": "header",
        "name": "Sec-WebSocket-Protocol",
        "description": "A short-lived single-use viewer ticket carried only as the xenoteer.ticket.<secret> WebSocket subprotocol alongside the exact binary subprotocol. Never place the secret in a URL or log it."
      }
    },
    "parameters": {
      "DesktopId": {
        "name": "desktop_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "DesktopGenerationQuery": {
        "name": "desktop_generation",
        "in": "query",
        "required": true,
        "schema": {"type": "string", "format": "uuid"}
      },
      "WindowReferenceToken": {
        "name": "reference_token",
        "in": "path",
        "required": true,
        "description": "Opaque principal- and exact-window-birth-bound token; it is not an XID.",
        "schema": {"type": "string", "minLength": 16, "maxLength": 1024, "pattern": "^[A-Za-z0-9_-]+$"}
      },
      "WindowCursorQuery": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "description": "Opaque cursor returned by the preceding list page.",
        "schema": {"type": "string", "minLength": 16, "maxLength": 1024, "pattern": "^[A-Za-z0-9_-]+$"}
      },
...
    },
    "headers": {
...
      "RetryAfter": {
        "description": "Current implementation emits 1 second for nonterminal command results and backoff responses.",
        "required": true,
        "schema": {"type": "integer", "minimum": 1, "example": 1}
      },
      "OptionalRetryAfter": {
        "description": "Present with value 1 while the returned command result is nonterminal; absent when terminal.",
        "schema": {"type": "integer", "minimum": 1, "example": 1}
      }
    },
...
    "responses": {
...
      "BadRequest": {
        "description": "Malformed JSON, invalid path/query/header, strict-shape failure, or unsupported protocol major.",
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
      "Unauthorized": {
        "description": "A valid Bearer credential is required.",
        "headers": {
          "WWW-Authenticate": {
            "schema": {"type": "string", "const": "Bearer realm=\"xenoteer\""}
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
      "Forbidden": {
        "description": "The principal lacks the required grant, or a browser WebSocket Origin is denied.",
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
      "NotFound": {
        "description": "The requested versioned resource does not exist or may not be disclosed.",
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
      "Conflict": {
        "description": "Stale generation, lease conflict, command ID/content conflict, or unsafe cancellation stage.",
        "headers": {
          "Retry-After": {"$ref": "#/components/headers/OptionalRetryAfter"}
        },
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"},
            "examples": {
              "staleReference": {"externalValue": "./examples/problem-stale-reference.json"}
            }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "The request body exceeds the configured ordinary-message byte limit.",
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
...
      "TooManyRequests": {
        "description": "A bounded admission slot, queue, rate limit, or quota is full.",
        "headers": {
          "Retry-After": {"$ref": "#/components/headers/RetryAfter"}
        },
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
      "CapabilityUnavailable": {
        "description": "The desktop or required subsystem is not ready.",
        "headers": {
          "Retry-After": {"$ref": "#/components/headers/RetryAfter"}
        },
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
      "RequestOutcomeUnknown": {
        "description": "The handler deadline elapsed after dispatch. This does not cancel accepted work.",
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"},
            "examples": {
              "exactCommandRetry": {"externalValue": "./examples/problem-request-outcome-unknown.json"},
              "noAutomaticRetry": {"externalValue": "./examples/problem-request-outcome-unknown-no-retry.json"}
            }
          }
        }
      }
    }
  }
}